_OF-rain-bot/requirements.txt
psycopg2-binary
asyncpg
websockets
pybit
python-binance
okx
httpx
tenacity
pyee
python-telegram-bot
python-bitget
gate_api